pyproject.toml
[tool.poetry]
name = "tornado-webserver"
version = "0.1.0"
description = ""
authors = ["Tatsuya Tomita <tat@seriousexplosion.net>"]
readme = "README.md"
packages = [{include = "tornado_webserver"}]

[tool.poetry.dependencies]
python = "^3.11"
tornado = "^6.3.2"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
